# [PERF] stock_account: optimize compute valuation