--- a/doc/cla/individual/paimonchan2.md
+++ b/doc/cla/individual/paimonchan2.md
@@ -1,0 +1,11 @@
+Indonesia, 2025-03-22
+
+I hereby agree to the terms of the Odoo Individual Contributor License
+Agreement v1.0.
+
+I declare that I am authorized and able to make this agreement and sign this
+declaration.
+
+Signed,
+
+Nazzun Hanif Ahsani n.h.shounen@gmail.com https://github.com/paimonchan