@robodoo 2025-03-22
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/203038.png)](https://mergebot.odoo.com/odoo/odoo/pull/203038)