@Whenrow 2025-04-08
@paimonchan Hello. Thank you for the PR. I see neither `product_id` nor `company_id` is indexed. Does those fields being indexed is not enough for you ?