@paimonchan 2025-04-08
> @paimonchan Hello. Thank you for the PR. I see neither `product_id` nor `company_id` is indexed. Does those fields being indexed is not enough for you ?

Hello @Whenrow , thank you for the reply.

Yes, I’ve tried indexing both of those columns, but the performance is still poor—more or less the same as before applying the indexes.

- I tried creating separate indexes on each column individually.
- I also tried using a composite index on both product_id and company_id.