@Whenrow 2025-04-11
Hi @paimonchan. I play a bit with your patch. I don't think `id` is necessary to be in the index. And I also get better results with `product_id` being first element. 
![image](https://github.com/user-attachments/assets/8d14c59e-aa62-4e88-9021-2a2bb3d5e08e)
Do you see the same on your side ? Thank you