@paimonchan 2025-04-11
Hi @Whenrow 
I tried the three options you suggested. Two of them did not include `id` in the index, and one did:
- Option 1 → Index on `(company_id, product_id, value, quantity)`
- Option 2 → Index on `(product_id, company_id, value, quantity)`
- Option 3 → Index on `(product_id, company_id, id, value, quantity)`

## Benchmark
| Option 1| Option 2 | Option 3|
|--------|--------|--------|
| 492 ms | 527 ms | 88 ms| 

## Result
Option 3 performs the best and is slightly better than my initial patch, but it still requires the `id` in the index.

## Option 1
### Composite Index
![(1) Screenshot 2025-04-11 195406](https://github.com/user-attachments/assets/9e7bc879-23c3-461a-9b0b-179d255b8614)
### Query Explain
```python
Finalize GroupAggregate  (cost=1000.56..199524.39 rows=560 width=80) (actual time=398.343..492.314 rows=1 loops=1)
  Group Key: product_id
  Buffers: shared hit=707593
  ->  Gather  (cost=1000.56..199496.39 rows=1120 width=80) (actual time=398.106..492.295 rows=3 loops=1)
        Workers Planned: 2
        Workers Launched: 2
        Buffers: shared hit=707593
        ->  Partial GroupAggregate  (cost=0.56..198384.39 rows=560 width=80) (actual time=395.053..395.054 rows=1 loops=3)
              Group Key: product_id
              Buffers: shared hit=707593
              ->  Parallel Index Scan using stock_valuation_company_product_index on stock_valuation_layer  
                  (cost=0.56..194521.64 rows=308348 width=20) 
                  (actual time=0.044..341.827 rows=247490 loops=3)
                    Index Cond: ((company_id = 1) AND (product_id = 140))
                    Buffers: shared hit=707593
Planning Time: 0.085 ms
Execution Time: 492.349 ms
``` 
## Option 2
### Composite Index
![(2) Screenshot 2025-04-11 200334](https://github.com/user-attachments/assets/df2f61a7-f3e6-4c2a-8abc-d8799ea97f19)
### Query Explain
```python
Finalize GroupAggregate  (cost=1000.56..402160.97 rows=560 width=80) (actual time=430.907..527.599 rows=1 loops=1)
  Group Key: product_id
  Buffers: shared hit=707599
  ->  Gather  (cost=1000.56..402132.97 rows=1120 width=80) (actual time=430.621..527.582 rows=3 loops=1)
        Workers Planned: 2
        Workers Launched: 2
        Buffers: shared hit=707599
        ->  Partial GroupAggregate  (cost=0.56..401020.97 rows=560 width=80) (actual time=427.539..427.540 rows=1 loops=3)
              Group Key: product_id
              Buffers: shared hit=707599
              ->  Parallel Index Scan using stock_valuation_company_product_index on stock_valuation_layer  
                  (cost=0.56..397158.22 rows=308348 width=20) 
                  (actual time=0.048..373.302 rows=247490 loops=3)
                    Index Cond: ((product_id = 140) AND (company_id = 1))
                    Buffers: shared hit=707599
Planning:
  Buffers: shared hit=30
Planning Time: 0.189 ms
Execution Time: 527.637 ms
``` 
## Option 3
### Composite Index
![(3) Screenshot 2025-04-11 201704](https://github.com/user-attachments/assets/f7d21b3a-e9ab-45e6-adfd-c0c4762a64b7)
### Query Explain
```python
Finalize GroupAggregate  (cost=1000.56..25246.11 rows=560 width=80) (actual time=80.996..88.120 rows=1 loops=1)
  Group Key: product_id
  Buffers: shared hit=19994
  ->  Gather  (cost=1000.56..25218.11 rows=1120 width=80) (actual time=80.872..88.106 rows=3 loops=1)
        Workers Planned: 2
        Workers Launched: 2
        Buffers: shared hit=19994
        ->  Partial GroupAggregate  (cost=0.56..24106.11 rows=560 width=80) (actual time=77.583..77.584 rows=1 loops=3)
              Group Key: product_id
              Buffers: shared hit=19994
              ->  Parallel Index Only Scan using stock_valuation_company_product_index on stock_valuation_layer  
                  (cost=0.56..20243.36 rows=308348 width=20) 
                  (actual time=0.044..36.614 rows=247490 loops=3)
                    Index Cond: ((product_id = 140) AND (company_id = 1))
                    Heap Fetches: 9201
                    Buffers: shared hit=19994
Planning Time: 0.081 ms
Execution Time: 88.151 ms
```