@paimonchan 2025-04-17
> @paimonchan Ok so let's keep `(product_id, company_id, id, value, quantity)`. thank you

@Whenrow Got it !
The changes are already in place. Thank you